requirements.txt
Pillow>=9.0.0
keyboard>=0.13.5
python-dotenv>=0.19.2